# client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ZPI API",
    "version": "all"
  },
  "paths": {
    "/api/assets": {
      "get": {
        "tags": ["Assets"],
        "operationId": "GetAllAssets",
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/asset-values/{assetName}": {
      "get": {
        "tags": ["AssetValues"],
        "operationId": "SearchAssetValuesHistory",
        "parameters": [
          {
            "name": "assetName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetValueDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/asset-values/{assetIdentifier}": {
      "post": {
        "tags": ["AssetValues"],
        "operationId": "PatchAssetValue",
        "parameters": [
          {
            "name": "assetIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAssetValueDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/users/me/assets": {
      "get": {
        "tags": ["UserAssets"],
        "operationId": "GetAllUserAssets",
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserAssetDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["UserAssets"],
        "operationId": "PatchUserAssets",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PatchUserAssetsDto"
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserAssetDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me/assets/{assetName}": {
      "delete": {
        "tags": ["UserAssets"],
        "operationId": "DeleteUserAsset",
        "parameters": [
          {
            "name": "assetName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserAssetDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me/preferences": {
      "put": {
        "tags": ["UserPreferences"],
        "operationId": "UpdateUserPreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserPreferencesDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferencesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["UserPreferences"],
        "operationId": "GetUserPreferences",
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferencesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/users": {
      "post": {
        "tags": ["Users"],
        "operationId": "AddUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUserDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddAssetValueDto": {
        "required": ["timeStamp", "value"],
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "format": "double"
          },
          "timeStamp": {
            "type": "string",
            "format": "date-time",
            "example": "2019-12-10T11:06:50.466318+01:00"
          }
        },
        "additionalProperties": false
      },
      "AddUserDto": {
        "required": ["userId"],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AssetDto": {
        "required": ["category", "friendlyName", "name", "symbol"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "friendlyName": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "symbol": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AssetValueDto": {
        "required": ["assetIdentifier", "timeStamp", "value"],
        "type": "object",
        "properties": {
          "assetIdentifier": {
            "type": "string"
          },
          "value": {
            "type": "number",
            "format": "double"
          },
          "timeStamp": {
            "type": "string",
            "format": "date-time",
            "example": "2019-12-10T11:06:50.466318+01:00"
          }
        },
        "additionalProperties": false
      },
      "PatchUserAssetsDto": {
        "required": ["assetName", "type", "value"],
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "format": "double"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "assetName": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "TransactionType": {
        "enum": ["Add", "Subtract", "Update"],
        "type": "string"
      },
      "UpdateUserPreferencesDto": {
        "required": ["alertsOnEmail", "preferenceCurrency", "weeklyReports"],
        "type": "object",
        "properties": {
          "preferenceCurrency": {
            "type": "string"
          },
          "weeklyReports": {
            "type": "boolean"
          },
          "alertsOnEmail": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UserAssetDto": {
        "required": ["asset", "originValue", "userCurrencyValue"],
        "type": "object",
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetDto"
          },
          "originValue": {
            "type": "number",
            "format": "double"
          },
          "userCurrencyValue": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "UserPreferencesDto": {
        "required": ["alertsOnEmail", "preferenceCurrency", "weeklyReports"],
        "type": "object",
        "properties": {
          "preferenceCurrency": {
            "type": "string"
          },
          "weeklyReports": {
            "type": "boolean"
          },
          "alertsOnEmail": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "BearerJWT": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerJWT": []
    }
  ]
}